Vagrant box is good, root login builds.

+++ bootstrap.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
-# bootstrap.sh
+# shell_provisioner.sh
 #
 # This script preps Ubuntu for a Packer build.
 
-EDITION=${EDITION:-server}
+# Default some ENV variable settings.
+COMPACT=${COMPACT:-"false"}
+DISABLE_ROOT=${DISABLE_ROOT:-"false"}
+INSTALL_SALT=${INSTALL_SALT:-"false"}
+SETUP_VAGRANT=${SETUP_VAGRANT:-"false"}
 
 # ==============================================================================
 # FUNCTION DEFINITIONS
@@ -27,33 +31,37 @@ clean_up() {
 # Place and erase zeros from empty space on drive. (Improves compression)
 compact() {
   echo "compact() function called"
-  sudo dd if=/dev/zero of=/EMPTY bs=1M
-  sudo rm -rf /EMPTY
+  if [ "$COMPACT" = "true" ]; then
+    sudo dd if=/dev/zero of=/EMPTY bs=1M
+    sudo rm -rf /EMPTY
+  else
+    echo "Skipping compact."
+  fi
 }
 
-# Fetch updates from /etc/apt/sources.list package indexes, then upgrade packages.
-upgrade() {
-  echo "upgrade() function called"
-  sudo apt-get update -y
-  sudo apt-get upgrade -y
+disable_root() {
+  echo "disable_root() function called"
+  if [ "$DISABLE_ROOT" = "true" ]; then
+    # Remove the root account SSH key.
+    sudo rm -rf /root/.ssh/
+    # Expire the root account. (deactivate)
+    sudo passwd -dl root
+  else
+    echo "ALERT: root account left active!!!"
+  fi
 }
 
 # Tested on Ubuntu 14.04
-# The latest packages for Ubuntu are published in the saltstack PPA.
 install_salt() {
   echo "install_salt() function called"
-  if [ -z $NO_SALT ]; then
+  if [ "$INSTALL_SALT" = "true" ]; then
     echo "Installing Salt..."
-    # Add official Salt PPA and install
-    sudo add-apt-repository -y ppa:saltstack/salt
-    sudo apt-get update
-    sudo apt-get install -y salt-master
-    sudo apt-get install -y salt-minion
-    # Setup firewall for Salt-master ports 4505, 4506.
-    # Test with: nmap -sS -q -p 4505-4506 salt.master.ip.addr
-    sudo ufw app update Salt
-    sudo ufw allow Salt
-    # Ensure services are running.
+    wget -O - https://bootstrap.saltstack.com | sudo sh -s -- $SALT_BOOTSTRAP_OPTIONS
+    # Setup firewall for default salt-master (4505), and salt-minion (4506) ports.
+    # Test with: nmap -sS -q -p 4505-4506 localhost
+    # sudo ufw app update Salt
+    # sudo ufw allow Salt
+    # Ensure services recognize any configurations.
     sudo service salt-master restart
     sudo service salt-minion restart
   else
@@ -61,29 +69,29 @@ install_salt() {
   fi
 }
 
-# Parse argruments in $@, either --server or --desktop must be passed.
+# Parse arguments in $@
 parse_options() {
   while :
   do
     case $1 in
-      --help)
+      -h | --help)
         show_usage
         exit 0
         ;;
-      --desktop)
-        EDITION=desktop
+      --compact)
+        COMPACT="true"
         shift
         ;;
-      --server)
-        EDITION=server
+      --disable-root)
+        DISABLE_ROOT="true"
         shift
         ;;
-      --no-salt)
-        NO_SALT=true
+      --install-salt)
+        INSTALL_SALT="true"
         shift
         ;;
-      --no-vagrant)
-        NO_VAGRANT=true
+      --setup-vagrant)
+        SETUP_VAGRANT="true"
         shift
         ;;
       ?*) # Invalid option.
@@ -98,6 +106,38 @@ parse_options() {
   done
 }
 
+# Ensure the vagrant user is ready for immediate usage by vagrant, then exit.
+setup_vagrant() {
+  echo "setup_vagrant() function called"
+  if [ "$SETUP_VAGRANT" = "true" ]; then
+    echo "Setting up Vagrant..."
+    # Create the vagrant user
+    sudo useradd --password '$6$h9l5Mhzb$dTlqr6wzzu9YrSECBHMhxLvUjcCbXu.6hYQS8o3dxlct3HhqegOJJ2gtYeRdkUCuInXe2ib/kXJF2Rji4JmGl.' \
+      --user-group \
+      --comment "Vagrant User" \
+      --create-home \
+      --shell /bin/bash \
+      vagrant
+    # Give password-less sudo priveledges, ensure privs are good.
+    echo "vagrant ALL=(ALL) NOPASSWD:ALL" | sudo tee /etc/sudoers.d/vagrant
+    sudo chown root:root /etc/sudoers.d/vagrant
+    sudo chmod 0440 /etc/sudoers.d/vagrant
+    # Setup ssh for the Vagrant User.
+    sudo mkdir -p --mode=0700 /home/vagrant/.ssh
+    sudo wget -qO /home/vagrant/.ssh/authorized_keys https://raw.github.com/mitchellh/vagrant/master/keys/vagrant.pub
+    sudo chmod 0600 /home/vagrant/.ssh/authorized_keys || { echo "~/.ssh/authorized_keys error!!!"; exit 1; }
+    sudo chown -R vagrant:vagrant /home/vagrant/.ssh
+    sudo echo "UseDNS no" >> /etc/ssh/sshd_config
+    # Install VB guest additions, build-essential is preseeded.
+    sudo apt-get install -y build-essential linux-headers-$(uname -r) dkms
+    sudo mount -o loop VBoxGuestAdditions.iso /media/cdrom
+    sudo sh /media/cdrom/VBoxLinuxAdditions.run
+    sudo umount /media/cdrom
+  else
+    echo "Skipping Vagrant setup."
+  fi
+}
+
 # Show usage info.
 show_usage() {
   cat <<EOF
@@ -105,43 +145,25 @@ show_usage() {
 Usage: $0 [--options]
 
   Defaults:
-    Server edition
-    Salt installed
-    Vagrant setup
+    Compact NOT performed
+    Salt NOT installed
+    Vagrant NOT setup
 
 Options:
+  --compact           Improve disk compression capability.
+  --disable-root      Expire the root account (de-activate)
   --help              Show this help menu.
-  --desktop           Bootstrap Ubuntu Desktop edition.
-  --server            Bootstrap Ubuntu Server edition.
-  --no-salt           Don't install Salt.
-  --no-vagrant        Don't setup Vagrant User.
+  --install-salt      Install salt master, minion, & ufw profile.
+  --setup-vagrant     Setup machine for Vagrant User, then exit.
 
 EOF
 }
 
-# Ensure the vagrant user is ready for immediate usage by vagrant.
-vagrant_user() {
-  echo "vagrant_user() function called"
-  if [ -z $NO_VAGRANT ]; then
-    echo "Setting up Vagrant User..."
-    # Install VB guest additions, build-essential is preseeded.
-    sudo apt-get install -y build-essential linux-headers-$(uname -r) dkms
-    sudo mount -o loop VBoxGuestAdditions.iso /media/cdrom
-    sudo sh /media/cdrom/VBoxLinuxAdditions.run
-    sudo umount /media/cdrom
-    # Allow passwordless sudo
-    echo 'vagrant ALL=(ALL) NOPASSWD:ALL' > /tmp/vagrant
-    sudo chown root:root /tmp/vagrant
-    sudo chmod 0440 /tmp/vagrant
-    sudo mv /tmp/vagrant /etc/sudoers.d/vagrant
-    # Setup ssh for the Vagrant User.
-    sudo mkdir -p --mode=0700 ~/.ssh
-    sudo wget -qO ~/.ssh/authorized_keys https://raw.github.com/mitchellh/vagrant/master/keys/vagrant.pub
-    sudo chmod 0600 ~/.ssh/authorized_keys || { echo "~/.ssh/authorized_keys error!!!"; exit 1; }
-    sudo chown -R vagrant:vagrant ~/.ssh
-  else
-    echo "Skipping Vagrant User setup."
-  fi
+# Fetch updates from /etc/apt/sources.list package indexes, then upgrade packages.
+upgrade() {
+  echo "upgrade() function called"
+  sudo apt-get update
+  sudo apt-get dist-upgrade -y
 }
 
 # ==============================================================================
@@ -151,13 +173,13 @@ vagrant_user() {
 # Get the command line options that were passed.
 parse_options $@
 
+# Setup the box for a default 'vagrant' user.
+setup_vagrant
+
 # Upgrade the selected updates.
 upgrade
 
-# Setup the box for a default 'vagrant' user.
-vagrant_user
-
-# Install the salt provisioner.
+# Install Salt (optional)
 install_salt
 
 # Clean up caches and old config files left around, manage the packagelist.
@@ -165,4 +187,7 @@ clean_up
 
 # Prep filesystem for better compression on VM.
 compact
+
+# Expire the root account.
+disable_root
